Split workspace memory parity gates
